more clean and seprated the codes
